# lab3/server/src/main/java/com/iit/aiposizi/generated/api/OfficesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.iit.aiposizi.generated.api;

import com.iit.aiposizi.generated.model.Office;
import java.util.UUID;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2020-05-17T23:29:22.869419+03:00[Europe/Minsk]")

@Validated
@Api(value = "offices", description = "the offices API")
public interface OfficesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /offices
     *
     * @param addressId  (required)
     * @param companyName  (required)
     * @return Created (status code 201)
     *         or Bad request (status code 400)
     *         or Not Found (status code 404)
     */
    @ApiOperation(value = "", nickname = "createOffice", notes = "", response = Office.class, tags={ "offices", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Created", response = Office.class),
        @ApiResponse(code = 400, message = "Bad request"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/offices",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    default ResponseEntity<Office> createOffice(@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "addressId", required = true) UUID addressId,@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "companyName", required = true) String companyName) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"address\" : \"address\", \"companyName\" : \"companyName\", \"roomsNumber\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /offices/{id}
     *
     * @param id  (required)
     * @return No content (status code 204)
     */
    @ApiOperation(value = "", nickname = "deleteOffice", notes = "", tags={ "offices", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "No content") })
    @RequestMapping(value = "/offices/{id}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deleteOffice(@ApiParam(value = "",required=true) @PathVariable("id") UUID id) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /offices/all
     *
     * @return OK (status code 200)
     */
    @ApiOperation(value = "", nickname = "getAllOffices", notes = "", response = Office.class, responseContainer = "List", tags={ "offices", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Office.class, responseContainer = "List") })
    @RequestMapping(value = "/offices/all",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Office>> getAllOffices() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"address\" : \"address\", \"companyName\" : \"companyName\", \"roomsNumber\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /offices/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Bad request (status code 400)
     *         or Not Found (status code 404)
     */
    @ApiOperation(value = "", nickname = "getOfficeById", notes = "", response = Office.class, tags={ "offices", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Office.class),
        @ApiResponse(code = 400, message = "Bad request"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/offices/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Office> getOfficeById(@ApiParam(value = "",required=true) @PathVariable("id") UUID id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"address\" : \"address\", \"companyName\" : \"companyName\", \"roomsNumber\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}