# lab3/server/src/main/java/com/iit/aiposizi/generated/api/PlacesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.iit.aiposizi.generated.api;

import com.iit.aiposizi.generated.model.Place;
import java.util.UUID;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2020-05-17T23:29:22.869419+03:00[Europe/Minsk]")

@Validated
@Api(value = "places", description = "the places API")
public interface PlacesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /places
     *
     * @param roomId  (required)
     * @param number  (required)
     * @return Created (status code 201)
     *         or Not Found (status code 404)
     */
    @ApiOperation(value = "", nickname = "createPlace", notes = "", response = Place.class, tags={ "places", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Created", response = Place.class),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/places",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    default ResponseEntity<Place> createPlace(@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "roomId", required = true) UUID roomId,@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "number", required = true) Integer number) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"number\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"employee\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"speciality\" : \"speciality\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"room\" : { \"totalPlaces\" : 6, \"number\" : 5, \"freePlaces\" : 1, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"office\" : { \"address\" : \"address\", \"companyName\" : \"companyName\", \"roomsNumber\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /places/{id}
     *
     * @param id  (required)
     * @return No Content (status code 204)
     *         or Bad request (status code 400)
     */
    @ApiOperation(value = "", nickname = "deletePlace", notes = "", tags={ "places", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "No Content"),
        @ApiResponse(code = 400, message = "Bad request") })
    @RequestMapping(value = "/places/{id}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deletePlace(@ApiParam(value = "",required=true) @PathVariable("id") UUID id) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /places/all
     *
     * @param offices  (optional, default to new ArrayList&lt;&gt;())
     * @param rooms  (optional, default to new ArrayList&lt;&gt;())
     * @return OK (status code 200)
     *         or Bad request (status code 400)
     */
    @ApiOperation(value = "", nickname = "getAllPlaces", notes = "", response = Place.class, responseContainer = "List", tags={ "places", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Place.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Bad request") })
    @RequestMapping(value = "/places/all",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Place>> getAllPlaces(@ApiParam(value = "") @Valid @RequestParam(value = "offices", required = false) List<String> offices,@ApiParam(value = "") @Valid @RequestParam(value = "rooms", required = false) List<Integer> rooms) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"number\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"employee\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"speciality\" : \"speciality\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"room\" : { \"totalPlaces\" : 6, \"number\" : 5, \"freePlaces\" : 1, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"office\" : { \"address\" : \"address\", \"companyName\" : \"companyName\", \"roomsNumber\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /places/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Bad request (status code 400)
     *         or Not Found (status code 404)
     */
    @ApiOperation(value = "", nickname = "getPlaceById", notes = "", response = Place.class, tags={ "places", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Place.class),
        @ApiResponse(code = 400, message = "Bad request"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/places/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Place> getPlaceById(@ApiParam(value = "",required=true) @PathVariable("id") UUID id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"number\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"employee\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"speciality\" : \"speciality\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"room\" : { \"totalPlaces\" : 6, \"number\" : 5, \"freePlaces\" : 1, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"office\" : { \"address\" : \"address\", \"companyName\" : \"companyName\", \"roomsNumber\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}