# lab3/server/src/main/resources/openapi/server.yml
openapi: 3.0.0
info:
  version: "1.0"
  title: Api Documentation
servers:
  - url: /api/v1
paths:
...
  /offices/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags:
        - offices
      operationId: getOfficeById
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
        400:
          description: Bad request
        404:
          description: Not Found
      deprecated: false
    delete:
      tags:
        - offices
      operationId: deleteOffice
      responses:
        204:
          description: No content
      deprecated: false
  /offices:
    post:
      tags:
        - offices
      operationId: createOffice
      parameters:
        - name: addressId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: companyName
          in: query
          required: true
          schema:
            type: string
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
        400:
          description: Bad request
        404:
          description: Not Found
      deprecated: false

  # ========================================
  # ===== Places Api
  # ========================================
  /places/all:
    get:
      tags:
        - places
      operationId: getAllPlaces
      parameters:
        - $ref: '#/components/parameters/Offices'
        - $ref: '#/components/parameters/Rooms'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
        400:
          description: Bad request
      deprecated: false
  /places/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags:
        - places
      operationId: getPlaceById
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        400:
          description: Bad request
        404:
          description: Not Found
      deprecated: false
    delete:
      tags:
        - places
      operationId: deletePlace
      responses:
        204:
          description: No Content
        400:
          description: Bad request
      deprecated: false
  /places:
    post:
      tags:
        - places
      operationId: createPlace
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Number'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        404:
          description: Not Found
      deprecated: false

  # ========================================
  # ===== Rooms Api
  # ========================================
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        speciality:
          type: string
...
    Office:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        companyName:
          type: string
        roomsNumber:
          type: integer
          format: int32
    Place:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee:
          $ref: "#/components/schemas/Employee"
        number:
          type: integer
          format: int32
        room:
          $ref: "#/components/schemas/Room"
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        totalPlaces:
          type: integer
          format: int32
        freePlaces:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        office:
          $ref: "#/components/schemas/Office"

  # ========================================
  # ===== Parameters
  # ========================================
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Offices:
      name: offices
      in: query
      required: false
      explode: true
      schema:
        default: []
        type: array
        items:
          type: string
    Number:
      name: number
      in: query
      required: true
      schema:
        type: integer
        format: int32
    Rooms:
      name: rooms
      in: query
      required: false
      explode: true
      schema:
        default: []
        type: array
        items:
          type: integer
          format: int32